float fix, add near filter validation, fix nested

+++ src/main/java/ru/smartel/strike/dto/response/event/EventDetailDTO.java
@@ -14,8 +14,8 @@ import java.util.Optional;
 
 public class EventDetailDTO extends PostDetailDTO {
 
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
     private long conflictId;
     private Long eventStatusId;
     private Long eventTypeId;
@@ -41,19 +41,19 @@ public class EventDetailDTO extends PostDetailDTO {
         return instance;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
